Renamed the crate to shared_memory.
This new name is more representative. The inital idea for this crate was
to provide File like operations that were backed by shared memory
instead of a file on disk. As the library evolved, it became clear that
accessing the shared memory through regular rust variables wrapped in a
lock would be much better than using Read/Write/Seek/etc...